Replace Category with Kind
Category was meant to describe different document types, but it was confusing and
can be mistaken as category within our docs (i.e., topics).
Kind should be a better name for the intention and we're going to add a new type
Topic to describe the main subject of the document.