Use Shiki to syntax highlight code blocks
Benefits of Shiki (https://shiki.matsu.io/):

- Accurate: uses TextMate grammar
- Pretty: uses VSCode theme
- Language Support: can extend by loading TextMate grammar

The appearance of code blocks is important for technical docs, and we
also support some niche languages.

Syntax highlighting is done at build time. A button to copy the code is
still supported. Decoration syntax from Docusaurus is not supported
at the moment.